fixed audio only playing in the right ear, fixed audio dropoff
The right audio channel bug was caused by passing a 1 to panStereo instead of a 0. However, this made me realize that the sound never "drops off" so to speak, because panStereo isn't truly 3D.

I set group.audio.rolloffMode to AudioRolloffMode to Logarithmic, because Linear had a choppy distance transition. dopplerLevel made things sound too distorted, so I set it to 0 (disabled.) I set SpatialBlend to 1, this is what the vanilla ModuleEngine uses and for good reason - it allows for 3D audio.